Move assetpack out of __init__
